+++ lib/droplet_kit/client.rb
@@ -23,6 +23,8 @@ module DropletKit
         cdns: CDNResource,
         certificates: CertificateResource,
         droplets: DropletResource,
+        kubernetes_clusters: KubernetesClusterResource,
+        kubernetes_options: KubernetesOptionsResource,
         domains: DomainResource,
         domain_records: DomainRecordResource,
         droplet_actions: DropletActionResource,